backend/requirements.txt
# Backend шаардлагатай package-үүд

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# File handling
aiofiles==23.2.1
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6